chore: separate methods to utils
